+++ src/main/java/com/demo/entrepreneur/entity/User.java
@@ -1,4 +1,4 @@
-package com.demo.entrepreneur.model;
+package com.demo.entrepreneur.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
